moved some .js files into /lib
