+++ docs/lib/component.js
