+++ docs/lib/factory.js
