+++ docs/lib/module.js
