+++ docs/lib/palette.js
