+++ docs/lib/progress_bar.js
